Refactor keyboard and display management: update I2C address, migrate key pins, enhance systems test functionality, and improve logging

+++ include/systems_test.h
@@ -45,7 +45,7 @@ private:
     };
     
     const char* getKeyName(char key);
-    uint8_t getKeyColor(char key);
+    uint16_t getKeyColor(char key);
 };
 
 #endif // STike_SYSTEM_TEST
